Added tabs for the web page
