+++ PatisserieProject/PatisserieFrontEnd/src/utilities/enum.ts
@@ -1,0 +1,10 @@
+enum productTypes {
+    macaroons = 'Macaroons',
+    fudge = 'Fudge',
+    miniTart = 'Mini Tart',
+    celebrationCakes = 'Celebration Cakes',
+}
+
+export default {
+    productTypes
+}
